Add logger functions and move player functions to feature
